requirement.txt
torch==2.1.0
torchvision==0.17.1
gradio==3.44.0
Pillow==10.0.0
